api/requirements.txt
fastapi
asyncpg
pydantic
python-dotenv
uvicorn
prometheus_client
pydantic